fix a bug that will cause incorrect matching between detector and tracker

+++ src/ptl_reid/src/ptl_reid.py
@@ -152,9 +152,8 @@ class ReIDNode():
             id_most_common = Counter(ids).most_common()[0][0]
             id_most_common_frequency = Counter(ids).most_common()[0][1]
             # old object
-            if (id_most_common_frequency >= int(
-                    len(ids) * self.database.cfg.batch_ratio)) & (
-                        id_most_common != -1):
+            if (id_most_common_frequency >= len(ids) *
+                    self.database.cfg.batch_ratio) & (id_most_common != -1):
                 id = id_most_common
                 is_query_new = False
                 self.database.add_new_feat(feats, id)
